+++ temba/orgs/models.py
@@ -2036,6 +2036,9 @@ class Org(SmartModel):
         for log in self.http_logs.all():
             log.release()
 
+        for g in self.globals.all():
+            g.release()
+
         # delete our classifiers
         for classifier in self.classifiers.all():
             classifier.release()
@@ -2071,6 +2074,10 @@ class Org(SmartModel):
         self.api_tokens.all().delete()
         self.invitations.all().delete()
         self.credit_alerts.all().delete()
+        self.broadcast_set.all().delete()
+        self.schedules.all().delete()
+
+        # needs to come after deletion of msgs and broadcasts as those insert new counts
         self.system_labels.all().delete()
 
         # now what we've all been waiting for